Move generic settings definition functions to input_structs.jl
This allows them to be called in other modules in a setup function,
which can then be called in moment_kinetics_input.jl, avoiding an
unresolvable cyclic import.
